Added WebUI Variables
Something weird about Pterodactyl variables, Omegga config and quotes. Basically, Pterodactyl's yaml bools are 0 or 1. Omegga requires true or false. Was able to have Pterodactyl output "true" or "false" by changing the variable's validation to (string|in:true,false). But since true / false is a bool in yaml, Pterodactyl will surround it in quotes to make it a string. Omegga doesn't read the values since it's not a valid bool.

Was able to resolve the issue by using the "yq" package. yq reads and modifies yaml configs. Basically, in entrypoint.sh it'll read both variables (omegga.https and omegga.webui). Then set both to the same values. yq will see it as a bool and not surround it in quotes, removing the quotes from both variables.